requirements.txt
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pyOpenSSL==23.2.0
gunicorn==21.2.0
psycopg2-binary==2.9.9 